api notification

+++ app/Http/Controllers/Api/V1/CommentaryController.php
@@ -7,6 +7,7 @@ use App\Models\Comment;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 
 class CommentaryController extends Controller
 {
@@ -15,7 +16,7 @@ class CommentaryController extends Controller
      */
     public function index(Post $post)
     {
-       $posts =  $post->comments()->get();
+        $posts = $post->comments()->get();
 
         return response()->json($posts);
     }
@@ -23,43 +24,51 @@ class CommentaryController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request,Post $post)
+    public function store(Request $request, Post $post)
     {
+
         $user = Auth::user();
-        if(!$user){
+        if (!$user) {
             return response()->json(['error' => 'Unauthorized'], 401);
         }
 
-      $comments =   Comment::create([
+        $comments = $post->comments()->create([
             'text' => $request->input('text'),
-            'post_id' => $post->id,
-            'user_id' => $user->id,
+            'user_id' => $user->id
         ]);
-        return response()->json(['message' => 'success',$comments],200);
+        return response()->json(['message' => 'success', $comments], 200);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(Post $post)
     {
-        //
+
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request,Post $post,Comment $comment)
     {
-        //
+        abort_if($comment->post_id !== $post->id,403);
+        $this->authorize('update', $comment);
+
+        $comment->update([
+            'text' => $request->input('text'),
+        ]);
+        return response()->json(['message' => 'success'], 200);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Comment $comment)
+    public function destroy(Post $post,Comment $comment)
     {
+        abort_if($comment->post_id !== $post->id,403);
+        $this->authorize('delete', $comment);
         $comment->delete();
-            return response()->json(['message' => 'success']);
-  }
+        return response()->json(['message' => 'success']);
+    }
 }
